separate files
classes instead of prototype