@Eomm 2023-09-15
Did you run benchmarks?

Splitting into multiple files means 3 more accesses to the file system.