@sirenkovladd 2023-09-15
> I don't see how this would give us better maintainability.

My point is that it will be more convenient for people to read the library
because each file has its own area of responsibility
it will also be more convenient to read the class instead of functions and prototype
it would be good to replace everything to one style, since other files use classes

It seems to me that when everything is in one file and written in a relatively old format, it is very difficult to maintain