@sirenkovladd 2023-09-15
> Did you run benchmarks?

```
find-my-way on  better-maintainability-stage-1 is 📦 v7.6.2 via  on ☁️  (us-west-2) took 15s 
❯ node ./benchmark/bench.js
lookup root "/" route.................................... x 22,029,625 ops/sec ±1.33% (580 runs sampled)
lookup short static route................................ x 11,043,140 ops/sec ±1.38% (577 runs sampled)
lookup long static route................................. x 3,692,081 ops/sec ±0.98% (582 runs sampled)
lookup long static route (common prefix)................. x 2,490,525 ops/sec ±0.87% (577 runs sampled)
lookup short parametric route............................ x 8,019,354 ops/sec ±1.42% (573 runs sampled)
lookup long parametric route............................. x 4,263,475 ops/sec ±1.44% (572 runs sampled)
lookup short parametric route (encoded unoptimized)...... x 4,063,224 ops/sec ±1.83% (575 runs sampled)
lookup short parametric route (encoded optimized)........ x 3,295,552 ops/sec ±1.35% (589 runs sampled)
lookup parametric route with two short params............ x 4,000,458 ops/sec ±1.11% (576 runs sampled)
lookup multi-parametric route with two short params...... x 3,117,017 ops/sec ±2.15% (546 runs sampled)
lookup multi-parametric route with two short regex params x 3,370,851 ops/sec ±1.29% (576 runs sampled)
lookup long static + parametric route.................... x 2,322,848 ops/sec ±1.03% (589 runs sampled)
lookup short wildcard route.............................. x 8,701,069 ops/sec ±1.37% (583 runs sampled)
lookup long wildcard route............................... x 4,548,452 ops/sec ±3.68% (572 runs sampled)
lookup root route on constrained router.................. x 17,775,709 ops/sec ±2.02% (571 runs sampled)
lookup short static unconstraint route................... x 10,287,354 ops/sec ±1.18% (574 runs sampled)
lookup short static versioned route...................... x 8,388,208 ops/sec ±2.06% (583 runs sampled)
lookup short static constrained (version & host) route... x 7,282,033 ops/sec ±1.36% (573 runs sampled)
```

```
find-my-way on  main is 📦 v7.6.2 via  v20.6.0 on ☁️  (us-west-2) 
❯ node ./benchmark/bench.js
lookup root "/" route.................................... x 21,294,883 ops/sec ±1.20% (569 runs sampled)
lookup short static route................................ x 11,331,300 ops/sec ±1.30% (575 runs sampled)
lookup long static route................................. x 3,477,830 ops/sec ±1.41% (584 runs sampled)
lookup long static route (common prefix)................. x 2,157,526 ops/sec ±0.74% (585 runs sampled)
lookup short parametric route............................ x 8,022,611 ops/sec ±1.29% (577 runs sampled)
lookup long parametric route............................. x 4,543,553 ops/sec ±0.68% (585 runs sampled)
lookup short parametric route (encoded unoptimized)...... x 4,669,611 ops/sec ±1.13% (578 runs sampled)
lookup short parametric route (encoded optimized)........ x 3,085,747 ops/sec ±1.01% (573 runs sampled)
lookup parametric route with two short params............ x 4,314,215 ops/sec ±1.22% (584 runs sampled)
lookup multi-parametric route with two short params...... x 3,570,758 ops/sec ±1.32% (571 runs sampled)
lookup multi-parametric route with two short regex params x 4,071,093 ops/sec ±0.96% (576 runs sampled)
lookup long static + parametric route.................... x 2,141,539 ops/sec ±1.15% (578 runs sampled)
lookup short wildcard route.............................. x 9,164,402 ops/sec ±1.17% (575 runs sampled)
lookup long wildcard route............................... x 4,592,144 ops/sec ±1.22% (572 runs sampled)
lookup root route on constrained router.................. x 17,828,236 ops/sec ±1.13% (572 runs sampled)
lookup short static unconstraint route................... x 9,898,024 ops/sec ±1.38% (565 runs sampled)
lookup short static versioned route...................... x 6,427,963 ops/sec ±1.04% (584 runs sampled)
lookup short static constrained (version & host) route... x 5,399,976 ops/sec ±1.12% (583 runs sampled)
```