@sirenkovladd 2023-09-15
> Splitting into multiple files means 3 more accesses to the file system.

I'm sure it didn't affect the benchmark because the library is imported there 1 time

But if you're really worried about importing multiple files, you can add a build that will put everything into 1 file without hurting the developer experience