@mcollina 2023-09-15
Looks like there are some regressions.
I think it's better to keep this module as-is.