@sirenkovladd 2023-09-15
```
lookup root "/" route. . . . . . . . . . . . . . . . . .  PR is faster by 3.45 %
lookup short static route. . . . . . . . . . . . . . . .  Main is faster by 2.61 %
lookup long static route. . . . . . . . . . . . . . . . . PR is faster by 6.16 %
lookup long static route (common prefix). . . . . . . . . PR is faster by 15.43 %
lookup short parametric route. . . . . . . . . . . . . .  Main is faster by 0.04 %
lookup long parametric route. . . . . . . . . . . . . . . Main is faster by 6.57 %
lookup short parametric route (encoded unoptimized). . .  Main is faster by 14.92 %
lookup short parametric route (encoded optimized). . . .  PR is faster by 6.80 %
lookup parametric route with two short params. . . . . .  Main is faster by 7.84 %
lookup multi-parametric route with two short params. . .  Main is faster by 14.56 %
lookup multi-parametric route with two short regex params Main is faster by 20.77 %
lookup long static + parametric route. . . . . . . . . .  PR is faster by 8.47 %
lookup short wildcard route. . . . . . . . . . . . . . .  Main is faster by 5.33 %
lookup long wildcard route. . . . . . . . . . . . . . . . Main is faster by 0.96 %
lookup root route on constrained router. . . . . . . . .  Main is faster by 0.30 %
lookup short static unconstraint route. . . . . . . . . . PR is faster by 3.93 %
lookup short static versioned route. . . . . . . . . . .  PR is faster by 30.50 %
lookup short static constrained (version & host) route. . PR is faster by 34.85 %
```

I'm sure that you can't rely on this benchmark, since you can take `lookup short static versioned route` as an example

30% difference in favor of PR (when nothing logically changed)
PR - 8,388,208 ops/sec ~ 119 ns/ops
main - 6,427,963 ops/sec ~ 155 ns/ops

node is not accurate enough to see the difference in 20988 ns (difference * sampled) for 574 samples
after all, these are external factors of my computer and GC