@sirenkovladd 2023-09-15
I believe that such quick tests have values on a logarithmic scale